+++ src/scriptlets/set-attr.js
@@ -24,6 +24,7 @@ import {
  * - `value` — the value to assign to the attribute, defaults to ''. Possible values:
  *     - `''` - empty string
  *     - positive decimal integer `<= 32767`
+ *     - `true` / `false` in any case variation
  *
  * **Examples**
  * 1.  Set attribute by selector
@@ -50,17 +51,45 @@ import {
  *     <!-- after -->
  *     <a class="class" test-attribute>Some text</div>
  *     ```
+ * 3.  Set attribute value to `TRUE`
+ *     ```
+ *     example.org#%#//scriptlet('set-attr', 'div.class > a.class', 'test-attribute', 'TRUE')
+ *     ```
+ *
+ *     ```html
+ *     <!-- before  -->
+ *     <a class="class">Some text</div>
+ *
+ *     <!-- after -->
+ *     <a class="class" test-attribute="TRUE">Some text</div>
+ *     ```
+ * 4.  Set attribute value to `fAlse`
+ *     ```
+ *     example.org#%#//scriptlet('set-attr', 'div.class > a.class', 'test-attribute', 'fAlse')
+ *     ```
+ *
+ *     ```html
+ *     <!-- before  -->
+ *     <a class="class">Some text</div>
+ *
+ *     <!-- after -->
+ *     <a class="class" test-attribute="fAlse">Some text</div>
+ *     ```
  */
 /* eslint-enable max-len */
 export function setAttr(source, selector, attr, value = '') {
     if (!selector || !attr) {
         return;
     }
+
+    const allowedValues = ['true', 'false'];
+
     // Drop strings that cant be parsed into number, negative numbers and numbers below 32767
     if (value.length !== 0
         && (nativeIsNaN(parseInt(value, 10))
             || parseInt(value, 10) < 0
-            || parseInt(value, 10) > 32767)) {
+            || parseInt(value, 10) > 32767)
+        && !allowedValues.includes(value.toLowerCase())) {
         return;
     }